restrucuture of main app and root files and implementation of global style
