+++ src/application/index.ts
@@ -1,0 +1,2 @@
+export { default as Root } from './Root';
+export { default as App } from './App';
