+++ src/styles/index.ts
@@ -1,0 +1,1 @@
+export { GlobalStyle } from './global.styles';
